requirements.txt
python-dotenv
requests
bcrypt
flask

